Mentor Mac: Needs refinement
please refer on my comments below:
Make sure that the file names we are using are on lower case, and there are no spaces, if there are spaces, we can replace them with hyphens, e.g. from "Team Card 1", we need to convert it to "team-card-1", this goes along not only with the file names but also with the images we are using within the images/ folder.
The WebStudio shouldn't be included in the <li> tag nor the <ul> tag.
Instead of using an address tag for the "Contacts" section, we can use a <ul> tag instead and wrap the items in an <li> tag
We can use a single <h1> tag to wrap the "Effective Solutions for Your Business" and use a self closing <br /> tag to separe them into 2 lines.
The hero section within the main tag should be wrapped with a <section> tag instead of a <div> tag, in total, we should have 4 section tags within the <main> tag.
Instead of using the <p> tag within "Section 1" for the titles, we can instead use an <h3> tag, we usually just use <p> for paragraphs or descriptions in this case.
For the team section, we can use an <h3> tag for the name, and a <p> tag for the title. The rule of thumb is as mentioned in #6, for descriptions/paragraphs, it would be <p> tag in most cases, and for titles or text that should be highlighted like the name, we usually use headings, or the <h3> tag in this case 😉
In the <p> tag within the <footer> tag, we can use self close <br /> tags to separate the lines.